♻️  Extract `enricher/depsdev` types into own file
